requirements.txt
pyyaml
diagrams
